requirements.txt
# 智能聊天预测程序依赖库

# 基础库
python-dotenv>=1.0.0  # 环境变量管理

# UI相关
PyQt5>=5.15.9  # 图形界面库

# 微信数据获取
pywin32>=306  # Windows API接口
pillow>=10.0.0  # 图像处理
pyperclip>=1.8.2  # 剪贴板操作

# API交互
openai>=1.3.0  # DeepSeek API调用
requests>=2.31.0  # HTTP请求

# 数据处理
numpy>=1.24.0
pandas>=2.0.0